Added new `build` files to help parsing API responses.
The `build_site()` and `build_collunit()` functions are intended to parse site and collectionunit specific data from the API.  This would then get passed into a site object. 

To help with debugging (and because the functions are generally useful) I also added an `@export` tag to `cleanNULL()`, `testNull()` and `use_na()`.