Minor text additions to improve article clarity.
In response to @mhesselbarth's comments in Issue #37 I have edited the Summary and body of the article.

To address the question about the `set_*()` functions I have indicated that the intention is to support upload, but that that the functionality (on the Neotoma Database side) is not present yet:

> The `set_*()` functions are intended as a precursor to utilities to upload data directly to Neotoma, although this functionality is not yet available.

The question about how the package conforms to a `tidyverse` model is explained here:

> This package conforms to a `tidyverse` [@wickham2019tidyverse] approach for data management, with methods and data objects that are suited to piping using the `%>%` (or `|>`) pipe convention and with implementations for filtering and other common `dplyr` methods.  The package also now uses "long" `data.frames` and `tibbles` by default, using the `toWide()` function to transform data into "wide" tables for use with common ecological data packages such as `vegan`.  Data objects in the `neotoma2` package now more closely resemble the underlying data model within Neotoma (https://open.neotomadb.org/db_schema) than in the previous package.

Change also reflects the addition of a bibtex entry for the `vegan` package.