NRP-21
Function to get speleothems - implemented with get_downloads but it may be more sensible to use get_datasets.

API locally needs an extra step to start returning datasets and sites calls, to speed up the process, I changed to downloads so that I could see and implement the calls and append the data to the sites object.